requirements.txt
flask>=3.0
crawl4ai>=0.4.0
python-dotenv>=1.0
lxml>=5.0
